src/requirements.txt
Flask
pandas
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client
Flask-SQLAlchemy
Flask-Session
num2words
Markdown2
WeasyPrint
Jinja2